pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ip-tray"
version = "0.3.0"
description = "macOS menu bar app: show IP, speeds, totals with country flag"
authors = [{ name = "IP Tray Maintainers" }]
requires-python = ">=3.9"
readme = "README.md"
license = { text = "MIT" }
keywords = ["macOS", "menubar", "network", "tray", "rumps"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: MacOS :: MacOS X",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "rumps>=0.4.0",
  "psutil>=5.9.0",
  "requests>=2.31.0",
  "Pillow>=10.0.0",
]

[project.scripts]
ip-tray = "ip_tray.app:run"

[tool.setuptools.packages.find]
where = ["src"]
